Avances en el plugin Locker. Implementado tooltip de carpetas (#63).

+++ plugins/locker/views/elements/file_info.ctp
@@ -24,19 +24,21 @@
 				$file['id']
 			);
 			$title = $name = $file['name'];
-			$ext = '';
-			$name = explode('.', $name);
-			$plus = 0;
-			if (count($name)==1) {
-				$plus = 4;
-				$name = $title;
-			} else {
-				$ext = '.' . array_pop($name);
-				$name = implode('.', $name);
+			if ($wrap_names) {
+				$ext = '';
+				$name = explode('.', $name);
+				$plus = 0;
+				if (count($name)==1) {
+					$plus = 4;
+					$name = $title;
+				} else {
+					$ext = '.' . array_pop($name);
+					$name = implode('.', $name);
+				}
+				$name = str_replace(' ', '^', $name);
+				$name = wordwrap($text->truncate($name, 16 + $plus, '[...]'), 10, '<br />', true);
+				$name = str_replace('^', ' ', $name) . $ext;
 			}
-			$name = str_replace(' ', '^', $name);
-			$name = wordwrap($text->truncate($name, 16 + $plus, '[...]'), 10, '<br />', true);
-			$name = str_replace('^', ' ', $name) . $ext;
 			echo $html->link($name, $url,
 				array(
 					'class'	=> 'item document ' . $file_type,
